copy existing lines below .zshrc others comments into new .zshrc

+++ deploy.sh
@@ -1,3 +1,10 @@
 cp ./tmux.conf ~/.tmux.conf
 cp ./vimrc ~/.vimrc
-cp ./zshrc ~/.zshrc
+
+if [ -z "$(grep -n 'others' ~/.zshrc | cut -d: -f 1)" ]; then
+    cp ./zshrc ~/.zshrc
+else
+    cp ./zshrc ~/.zshrc.tmp
+    tail -n +"$(grep -n 'others' ~/.zshrc | cut -d: -f 1)" ~/.zshrc >> ~/.zshrc.tmp
+    mv ~/.zshrc.tmp ~/.zshrc
+fi
